增加springsecurity canaccess,动态管理权限

+++ hy_teacherServer/src/main/java/com/moxi/teacherServer/restApi/classRestApi.java
@@ -4,11 +4,12 @@ import com.github.xiaoymin.knife4j.annotations.ApiOperationSupport;
 import com.moxi.base.exception.ThrowableUtils;
 import com.moxi.base.validator.group.GetList;
 import com.moxi.base.validator.group.Insert;
-import com.moxi.teacherServer.annotation.authority.AuthorityVerify;
 import com.moxi.teacherServer.annotation.permissionLog.permissionLogVerify;
 import com.moxi.teacherServer.global.SysConf;
 import com.moxi.utils.ResultUtil;
 import com.moxi.utils.ServerInfo.Sys;
+import com.moxi.utils.StringUtils;
+import com.moxi.xo.global.MessageConf;
 import com.moxi.xo.service.ClassService;
 import com.moxi.xo.vo.ClassVo;
 import io.swagger.annotations.Api;
@@ -36,7 +37,7 @@ import java.util.Map;
  * @date 2020/9/17 15:59
  */
 @RestController
-@RequestMapping("/teacher/class")
+@RequestMapping("/class")
 @Api(value = "4.班级相关接口", tags = {"4.班级相关接口"})
 public class classRestApi {
     @Autowired
@@ -54,11 +55,7 @@ public class classRestApi {
 
 
 
-//    @permissionLogVerify(
-//            resourceType = SysConf.RESOURCE_CLASS,
-//            operationType = {SysConf.OPERATION_ADD,SysConf.OPERATION_DELETE,SysConf.OPERATION_EDIT,SysConf.OPERATION_GETLIST},
-//            operand = {SysConf.OPERAND_STU,SysConf.OPERAND_EXAM, SysConf.OPERAND_PERMISSION_GROUP}
-//            )
+
     @ApiOperation(value = "新增班级", notes = "新增班级", response = String.class)
     @ApiOperationSupport(ignoreParameters ={"uid","currentPage","pageSize","keyword"})
     @PostMapping("/add")
@@ -69,15 +66,17 @@ public class classRestApi {
 
     @ApiOperation(value = "更新班级信息", notes = "更新班级信息", response = String.class)
     @ApiOperationSupport(ignoreParameters ={"currentPage","teacherId","pageSize","keyword"})
-    @PostMapping("/edit")
-    public String edit(@Validated({Update.class}) @RequestBody ClassVo vo) {
+    @PostMapping("/{classId}/edit")
+    public String edit(@PathVariable String classId,@Validated({Update.class}) @RequestBody ClassVo vo) {
+        if(StringUtils.isEmpty(classId))return ResultUtil.result(SysConf.ERROR, MessageConf.PARAM_INCORRECT);
         return classService.edit(vo);
     }
 
 
     @ApiOperation(value = "删除班级", notes = "删除班级", response = String.class)
-    @DeleteMapping("/delete/{classId}")
+    @DeleteMapping("/{classId}/delete")
     public String delete(@PathVariable String classId){
+        if(StringUtils.isEmpty(classId))return ResultUtil.result(SysConf.ERROR, MessageConf.PARAM_INCORRECT);
         return classService.delete(classId);
     }
 }
